Znakomstvo s Python/lesson_5-Schitaiem_siekundy_v_Telegram/requirements.txt
python-telegram-bot == 12.4.1
pytimeparse == 1.1.8
python-dotenv == 0.11.0